Move the last file to the right place
