+++ docs/modules/esdbConfig/components/security/CertSan.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script>
-import {safe} from "../../../../.vuepress/theme/util/strings";
+import {safe} from "../../../common/strings";
 
 export default {
     name:     "CertSan",
